refactor: renombrar package para seguir con las normas de Java
