--- backend-springboot/src/main/java/com/hasbi/taskManager/TaskManagerApplication.java
@@ -1,13 +1,0 @@
-package com.hasbi.taskManager;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class TaskManagerApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(TaskManagerApplication.class, args);
-	}
-
-}
