--- backend-springboot/src/main/java/com/hasbi/taskManager/controller/ProjectController.java
@@ -1,44 +1,0 @@
-package com.hasbi.taskManager.controller;
-
-import com.hasbi.taskManager.dto.ProjectDto;
-import com.hasbi.taskManager.service.ProjectService;
-import lombok.AllArgsConstructor;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
-
-@RestController
-@RequestMapping("/projects")
-@AllArgsConstructor
-public class ProjectController {
-
-    private ProjectService projectService;
-
-    @PostMapping("/add")
-    public ResponseEntity<ProjectDto> createProject(@RequestBody ProjectDto projectDto) {
-        ProjectDto project = projectService.createProject(projectDto);
-        return ResponseEntity.ok(project);
-    }
-
-    @GetMapping("/all")
-    public ResponseEntity<List<ProjectDto>> getAllProjects() {
-        return ResponseEntity.ok(projectService.getAllProjects());
-    }
-
-    @GetMapping("/{id}")
-    public ResponseEntity<ProjectDto> getProjectById(@PathVariable Long id) {
-        return ResponseEntity.ok(projectService.getProjectById(id));
-    }
-
-    @PutMapping("/update/{id}")
-    public ResponseEntity<ProjectDto> updateProject(@PathVariable Long id, @RequestBody ProjectDto projectDto) {
-        return ResponseEntity.ok(projectService.updateProject(id, projectDto));
-    }
-
-    @DeleteMapping("/delete/{id}")
-    public ResponseEntity<Void> deleteProject(@PathVariable Long id) {
-        projectService.deleteProject(id);
-        return ResponseEntity.noContent().build();
-    }
-}
