--- backend-springboot/src/main/java/com/hasbi/taskManager/controller/TaskController.java
@@ -1,44 +1,0 @@
-package com.hasbi.taskManager.controller;
-
-import com.hasbi.taskManager.dto.TaskDto;
-import com.hasbi.taskManager.service.TaskService;
-import lombok.AllArgsConstructor;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
-
-@RestController
-@RequestMapping("/tasks")
-@AllArgsConstructor
-public class TaskController {
-
-    private TaskService taskService;
-
-    @PostMapping("/add")
-    public ResponseEntity<TaskDto> createTask(@RequestBody TaskDto taskDto) {
-        TaskDto created = taskService.createTask(taskDto);
-        return ResponseEntity.ok(created);
-    }
-
-    @GetMapping("/project/{projectId}")
-    public ResponseEntity<List<TaskDto>> getTasksByProjectId(@PathVariable Long projectId) {
-        return ResponseEntity.ok(taskService.getTasksByProjectId(projectId));
-    }
-
-    @GetMapping("/{id}")
-    public ResponseEntity<TaskDto> getTaskById(@PathVariable Long id) {
-        return ResponseEntity.ok(taskService.getTaskById(id));
-    }
-
-    @PutMapping("/update/{id}")
-    public ResponseEntity<TaskDto> updateTask(@PathVariable Long id, @RequestBody TaskDto taskDto) {
-        return ResponseEntity.ok(taskService.updateTask(id, taskDto));
-    }
-
-    @DeleteMapping("/delete/{id}")
-    public ResponseEntity<Void> deleteTask(@PathVariable Long id) {
-        taskService.deleteTask(id);
-        return ResponseEntity.noContent().build();
-    }
-}
