--- backend-springboot/src/main/java/com/hasbi/taskManager/entity/Task.java
@@ -1,31 +1,0 @@
-package com.hasbi.taskManager.entity;
-
-import com.hasbi.taskManager.enums.TaskPriority;
-import com.hasbi.taskManager.enums.TaskStatus;
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Data  
-@NoArgsConstructor
-@AllArgsConstructor
-public class Task {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    private String title;
-    private String description;
-
-    @Enumerated(EnumType.STRING)
-    private TaskStatus status;
-
-    @Enumerated(EnumType.STRING)
-    private TaskPriority priority;
-
-    @ManyToOne
-    @JoinColumn(name = "project_id")
-    private Project project;
-}
