--- backend-springboot/src/main/java/com/hasbi/taskManager/enums/TaskPriority.java
@@ -1,5 +1,0 @@
-package com.hasbi.taskManager.enums;
-
-public enum TaskPriority {
-    LOW, MEDIUM, HIGH
-}
