--- backend-springboot/src/main/java/com/hasbi/taskManager/enums/TaskStatus.java
@@ -1,5 +1,0 @@
-package com.hasbi.taskManager.enums;
-
-public enum TaskStatus {
-    TODO, IN_PROGRESS, DONE
-}
