--- backend-springboot/src/main/java/com/hasbi/taskManager/exception/ResourceNotFoundException.java
@@ -1,7 +1,0 @@
-package com.hasbi.taskManager.exception;
-
-public class ResourceNotFoundException extends RuntimeException {
-    public ResourceNotFoundException(String message) {
-        super(message);
-    }
-}
