--- backend-springboot/src/main/java/com/hasbi/taskManager/repository/ProjectRepository.java
@@ -1,8 +1,0 @@
-package com.hasbi.taskManager.repository;
-
-import com.hasbi.taskManager.entity.Project;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-
-public interface ProjectRepository extends JpaRepository<Project, Long> {
-}
