--- backend-springboot/src/main/java/com/hasbi/taskManager/repository/TaskRepository.java
@@ -1,10 +1,0 @@
-package com.hasbi.taskManager.repository;
-
-import com.hasbi.taskManager.entity.Task;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
-
-public interface TaskRepository extends JpaRepository<Task, Long> {
-    List<Task> findByProjectId(Long projectId);
-}
