--- backend-springboot/src/main/java/com/hasbi/taskManager/service/ProjectService.java
@@ -1,13 +1,0 @@
-package com.hasbi.taskManager.service;
-
-import com.hasbi.taskManager.dto.ProjectDto;
-
-import java.util.List;
-
-public interface ProjectService {
-    ProjectDto createProject(ProjectDto projectDto);
-    List<ProjectDto> getAllProjects();
-    ProjectDto getProjectById(Long id);
-    ProjectDto updateProject(Long id, ProjectDto projectDto);
-    void deleteProject(Long id);
-}
