+++ backend-springboot/src/main/java/com/hasbi/taskmanager/TaskManagerApplication.java
@@ -1,0 +1,13 @@
+package com.hasbi.taskmanager;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class TaskManagerApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(TaskManagerApplication.class, args);
+	}
+
+}
