+++ backend-springboot/src/main/java/com/hasbi/taskmanager/dto/ProjectDto.java
@@ -1,0 +1,16 @@
+package com.hasbi.taskmanager.dto;
+
+import jakarta.validation.constraints.NotBlank;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class ProjectDto {
+    private Long id;
+    @NotBlank(message = "Name is required")
+    private String name;
+    private String description;
+}
