+++ backend-springboot/src/main/java/com/hasbi/taskmanager/enums/TaskPriority.java
@@ -1,0 +1,5 @@
+package com.hasbi.taskmanager.enums;
+
+public enum TaskPriority {
+    LOW, MEDIUM, HIGH
+}
