+++ backend-springboot/src/main/java/com/hasbi/taskmanager/enums/TaskStatus.java
@@ -1,0 +1,5 @@
+package com.hasbi.taskmanager.enums;
+
+public enum TaskStatus {
+    TODO, IN_PROGRESS, DONE
+}
