+++ backend-springboot/src/main/java/com/hasbi/taskmanager/exception/ResourceNotFoundException.java
@@ -1,0 +1,7 @@
+package com.hasbi.taskmanager.exception;
+
+public class ResourceNotFoundException extends RuntimeException {
+    public ResourceNotFoundException(String message) {
+        super(message);
+    }
+}
