+++ backend-springboot/src/main/java/com/hasbi/taskmanager/mapper/TaskMapper.java
@@ -1,0 +1,15 @@
+package com.hasbi.taskmanager.mapper;
+
+import com.hasbi.taskmanager.dto.TaskDto;
+import com.hasbi.taskmanager.entity.Task;
+import org.mapstruct.Mapper;
+import org.mapstruct.Mapping;
+
+@Mapper(componentModel = "spring")
+public interface TaskMapper {
+    @Mapping(source = "project.id", target = "projectId")
+    TaskDto toDto(Task task);
+
+    @Mapping(source = "projectId", target = "project.id")
+    Task toEntity(TaskDto dto);
+}
