+++ backend-springboot/src/main/java/com/hasbi/taskmanager/repository/TaskRepository.java
@@ -1,0 +1,10 @@
+package com.hasbi.taskmanager.repository;
+
+import com.hasbi.taskmanager.entity.Task;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.List;
+
+public interface TaskRepository extends JpaRepository<Task, Long> {
+    List<Task> findByProjectId(Long projectId);
+}
