+++ backend-springboot/src/main/java/com/hasbi/taskmanager/service/ProjectService.java
@@ -1,0 +1,13 @@
+package com.hasbi.taskmanager.service;
+
+import com.hasbi.taskmanager.dto.ProjectDto;
+
+import java.util.List;
+
+public interface ProjectService {
+    ProjectDto createProject(ProjectDto projectDto);
+    List<ProjectDto> getAllProjects();
+    ProjectDto getProjectById(Long id);
+    ProjectDto updateProject(Long id, ProjectDto projectDto);
+    void deleteProject(Long id);
+}
