--- backend-springboot/src/test/java/com/hasbi/taskManager/TaskManagerApplicationTests.java
@@ -1,15 +1,0 @@
-package com.hasbi.taskManager;
-
-import org.junit.jupiter.api.Disabled;
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@Disabled("Desactivado temporalmente mientras se configura el contexto de pruebas")
-@SpringBootTest
-class TaskManagerApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
