+++ backend-springboot/src/test/java/com/hasbi/taskmanager/TaskManagerApplicationTests.java
@@ -1,0 +1,15 @@
+package com.hasbi.taskmanager;
+
+import org.junit.jupiter.api.Disabled;
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@Disabled("Desactivado temporalmente mientras se configura el contexto de pruebas")
+@SpringBootTest
+class TaskManagerApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
